backend/requirements.txt
python-socketio
aiohttp
pigpio
pytest-aiohttp
pytest-asyncio
pytest-cov